Tag consul services by fqdn of host node (mesos slave)
Since we want the local haproxy on each node to only load-balance across
containers running on that node, we need to tell consul how to limit
the node list. Using a tag looks like a good way, but it is a bit weird
because we can't use any dots in the tag (prolly because of the DNS
nature of consul).

Also this features a change to the Marathon HTTP event bus -- apparently
only the master does the HTTP notifying. No matter -- we can just stick
every node's FQDN in the connection string and make the master do a lot
of work. This has the downside that (because I think the list of
callbackUrls is stored in ZK) it tends to clutter up, and right now
there is no way to clean out slaves that have died. Restarting the whole
cluster seems to work, but that seems a tad dramatic.